Moved async function to another step of the generator
