+++ generators/app/index.js
@@ -11,6 +11,8 @@ const bundleInstall = require('./tasks/bundleInstall');
 const configureFastlane = require('./tasks/configureFastlane');
 const installPods = require('./tasks/installPods');
 const linkAppAssets = require('./tasks/linkAppAssets');
+const editBundleIdentifier = require('./tasks/editBundleIdentifier');
+const configureIosProject = require('./tasks/configureIosProject');
 
 class ReactNativeBootstrap extends Generator {
   constructor(args, opts) {
@@ -97,8 +99,10 @@ class ReactNativeBootstrap extends Generator {
   install() {
     return Promise.resolve()
       .then(() => bundleInstall.bind(this)())
+      .then(() => configureIosProject.bind(this)())
       .then(() => installPods.bind(this)())
       .then(() => linkAppAssets.bind(this)())
+      .then(() => editBundleIdentifier.bind(this)())
       .then(() => gitInitialization.bind(this)())
       .then(() => this.features.bitrise && bitriseInitialization.bind(this)());
   }
