+++ generators/app/tasks/editBundleIdentifier.js
@@ -1,4 +1,4 @@
-const runCommand = require('../runCommand');
+const runCommand = require('./runCommand');
 
 module.exports = function editBundleIdentifier() {
   return runCommand({
@@ -14,7 +14,7 @@ module.exports = function editBundleIdentifier() {
       ],
       { cwd: `${process.cwd()}/${this.projectName}/ios` }
     ],
-    loadingMessage: 'Updating bundle identifier...\n',
+    loadingMessage: 'Updating bundle identifier...',
     successMessage: 'Bundle identifier updated!',
     failureMessage: 'Bundle identifier update failed. Turn verbose mode on for detailed logging',
     context: this.options
