+++ generators/app/tasks/installPods.js
@@ -3,7 +3,7 @@ const runCommand = require('./runCommand');
 module.exports = function installPods() {
   return runCommand({
     command: ['pod', ['install', '--repo-update'], { cwd: `${process.cwd()}/${this.projectName}/ios` }],
-    loadingMessage: 'Updating and installing Pods...\n',
+    loadingMessage: 'Updating and installing Pods...',
     successMessage: 'Pods ready!',
     failureMessage: 'Pod install failed. Turn verbose mode on for detailed logging',
     context: this.options
