+++ generators/app/tasks/scriptIosConfig.rb
